Organize tests into unit tests
