+++ test/unit/crowdloan/crowdloanFactory.test.js
@@ -7,7 +7,6 @@ import {
 
 const DebtTokenFactory = artifacts.require("DebtTokenFactory");
 const CrowdloanFactory = artifacts.require("CrowdloanFactory");
-const Crowdloan = artifacts.require("Crowdloan");
 
 const loanParams = {
   debtToken: "0x89d24A6b4CcB1B6fAA2625fE562bDD9a23260359",
